requirements.txt
pandas>=1.5.0
ezdxf>=1.0.0
openpyxl>=3.1.0
matplotlib>=3.8.3
Pillow>=11.2.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.26.0 